+++ changelog.md
@@ -2,6 +2,7 @@
 
 - Improved label reordering
 - Fixed Crash using Chest Unlock Bypass
+- Removed "him" from the credits secret
 - Added **Gold Chest Bypass** to Chest Unlock Bypass
 - Added **Localization / Translations**!!
 -- The language can be changed in the **Config** tab of the mod menu